test(gateway): split node pairing diagnostics suite
